# Conflicts:
#	kernel/audio/audio.cpp

+++ utils/terminal/terminal.cpp
@@ -4,6 +4,10 @@
 
 Terminal::Terminal() : Console() {
     char_scale = 2;
+    put_string("> ");
+    prompt_length = 2;
+    draw_cursor();
+    flush(dctx);
 }
 
 void Terminal::update(){
@@ -17,6 +21,8 @@ void Terminal::end_command(){
     command_running = false;
     put_char('\r');
     put_char('\n');
+    put_string("> ");
+    prompt_length = 2;
     draw_cursor();
     flush(dctx);
     set_text_color(default_text_color);
@@ -72,18 +78,33 @@ const char** Terminal::parse_arguments(char *args, int *count){
 
 void Terminal::run_command(){
     const char* fullcmd = get_current_line();
-    const char* args = seek_to(fullcmd, ' ');
     
+    if (fullcmd[0] == '>' && fullcmd[1] == ' ') {
+        fullcmd += 2;
+    }
+    
+    const char* check = fullcmd;
+    while (*check == ' ' || *check == '\t') check++;
+    if (*check == '\0') { 
+        prompt_length = 2;
+        draw_cursor();
+        flush(dctx);
+        command_running = true;
+        return;
+    }
+
+    const char* args = seek_to(fullcmd, ' ');
+
     string cmd;
     int argc = 0;
-    const char** argv; 
+    const char** argv = nullptr;
     string args_copy = {};
-    
-    if (fullcmd == args){
+
+    if (*args == '\0'){
         cmd = string_from_literal(fullcmd);
-        argv = 0;
     } else {
-        cmd = string_from_literal_length(fullcmd, args - fullcmd - 1);
+        size_t cmd_len = (size_t)((args - fullcmd) - 1);
+        cmd = string_from_literal_length(fullcmd, cmd_len);
         args_copy = string_from_literal(args);
         argv = parse_arguments(args_copy.data, &argc);
     }
@@ -102,10 +123,10 @@ void Terminal::run_command(){
             free(s.data, s.mem_length);
         }
     }
-    
+
     free(cmd.data, cmd.mem_length);
     if (args_copy.mem_length) free(args_copy.data, args_copy.mem_length);
-    
+
     draw_cursor();
     flush(dctx);
     command_running = true;
@@ -135,7 +156,9 @@ void Terminal::handle_input(){
                 draw_cursor();
                 flush(dctx);
             } else if (key == KEY_BACKSPACE){
-                delete_last_char();
+                if (strlen(get_current_line(), 1024) > (uint32_t)prompt_length) {
+                    delete_last_char();
+                }
             }
         }
     }
